Refactor AuctionGovernanceWidget and AuctionView to enhance UI and data handling. Remove unused logging, update content structure, and implement image handling for auction items. Introduce formatSecondsToDhms utility for better time display. Adjust AuctionData to use Int for endTime and update widget layout dimensions for improved responsiveness.

+++ apps/mobileapp/android/app/src/main/java/com/mobileapp/widgets/models/AuctionData.kt
@@ -8,7 +8,7 @@ data class AuctionData(
     val id: Int,
     @SerialName("amount") val currentBid: Double,
     val bidder: String,
-    val endTime: Double,
+    val endTime: Int,
     val image: String,
     val duration: Int
 )
